support for the CHIRP Audio Trigger board
several tweaks and changes throughout the code here. I added blinky lights for user feedback, as well as some basic chirp sounds as the system syncs WAV files from SD card to flash memory (which can take a couple of minutes, so hopefully now the user is more aware that things are happening).